# spec/api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "meteroid",
    "description": "",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/coupons": {
      "get": {
        "tags": [
          "Coupons"
        ],
        "summary": "List coupons",
        "operationId": "list_coupons",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ALL",
                "ACTIVE",
                "INACTIVE",
                "ARCHIVED"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number (0-indexed)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of coupons",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CouponListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Coupons"
        ],
        "summary": "Create a coupon",
        "operationId": "create_coupon",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCouponRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Coupon created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Coupon"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/coupons/{coupon_id}": {
      "get": {
        "tags": [
          "Coupons"
        ],
        "summary": "Get coupon details",
        "operationId": "get_coupon",
        "parameters": [
          {
            "name": "coupon_id",
            "in": "path",
            "description": "Coupon ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CouponId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coupon details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Coupon"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Coupon not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Coupons"
        ],
        "summary": "Update a coupon",
        "operationId": "update_coupon",
        "parameters": [
          {
            "name": "coupon_id",
            "in": "path",
            "description": "Coupon ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CouponId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCouponRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Coupon updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Coupon"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Coupon not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/coupons/{coupon_id}/archive": {
      "post": {
        "tags": [
          "Coupons"
        ],
        "summary": "Archive a coupon",
        "operationId": "archive_coupon",
        "parameters": [
          {
            "name": "coupon_id",
            "in": "path",
            "description": "Coupon ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CouponId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Coupon archived"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Coupon not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/coupons/{coupon_id}/disable": {
      "post": {
        "tags": [
          "Coupons"
        ],
        "summary": "Disable a coupon",
        "operationId": "disable_coupon",
        "parameters": [
          {
            "name": "coupon_id",
            "in": "path",
            "description": "Coupon ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CouponId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Coupon disabled"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Coupon not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/coupons/{coupon_id}/enable": {
      "post": {
        "tags": [
          "Coupons"
        ],
        "summary": "Enable a coupon",
        "operationId": "enable_coupon",
        "parameters": [
          {
            "name": "coupon_id",
            "in": "path",
            "description": "Coupon ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CouponId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Coupon enabled"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Coupon not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/coupons/{coupon_id}/unarchive": {
      "post": {
        "tags": [
          "Coupons"
        ],
        "summary": "Unarchive a coupon",
        "operationId": "unarchive_coupon",
        "parameters": [
          {
            "name": "coupon_id",
            "in": "path",
            "description": "Coupon ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CouponId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Coupon unarchived"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Coupon not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Coupon": {
        "type": "object",
        "required": [
          "id",
          "code",
          "description",
          "discount",
          "reusable",
          "disabled"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "disabled": {
            "type": "boolean"
          },
          "discount": {
            "$ref": "#/components/schemas/CouponDiscount"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "id": {
            "$ref": "#/components/schemas/CouponId"
          },
          "recurring_value": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "redemption_limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "reusable": {
            "type": "boolean"
          }
        }
      },
      "CouponDiscount": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/PercentageDiscount"
          },
          {
            "$ref": "#/components/schemas/FixedDiscount"
          }
        ],
        "discriminator": {
          "propertyName": "discriminator",
          "mapping": {
            "FIXED": "#/components/schemas/FixedDiscount",
            "PERCENTAGE": "#/components/schemas/PercentageDiscount"
          }
        }
      },
      "CouponDiscountRest": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "percentage",
              "discount_type"
            ],
            "properties": {
              "discount_type": {
                "type": "string",
                "enum": [
                  "PERCENTAGE"
                ]
              },
              "percentage": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "currency",
              "amount",
              "discount_type"
            ],
            "properties": {
              "amount": {
                "type": "string"
              },
              "currency": {
                "type": "string"
              },
              "discount_type": {
                "type": "string",
                "enum": [
                  "FIXED"
                ]
              }
            }
          }
        ]
      },
...
      "CouponId": {
        "type": "string",
        "format": "MeteroidId",
        "examples": [
          "cou_7n42DGM5Tflk9n8mt7Fhc7"
        ]
      },
...
      "CouponListResponse": {
        "type": "object",
        "required": [
          "data",
          "pagination_meta"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Coupon"
            }
          },
          "pagination_meta": {
            "$ref": "#/components/schemas/PaginationResponse"
          }
        }
      },
...
      "CreateCouponRequest": {
        "type": "object",
        "required": [
          "code",
          "discount"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "discount": {
            "$ref": "#/components/schemas/CouponDiscountRest"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "plan_ids": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanId"
            }
          },
          "recurring_value": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "redemption_limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "reusable": {
            "type": "boolean"
          }
        }
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "BAD_REQUEST",
          "NOT_FOUND",
          "CONFLICT",
          "FORBIDDEN",
          "UNAUTHORIZED",
          "INTERNAL_SERVER_ERROR"
        ]
      },
...
      "FixedDiscount": {
        "type": "object",
        "required": [
          "currency",
          "amount"
        ],
        "properties": {
          "amount": {
            "type": "string",
            "format": "decimal"
          },
          "currency": {
            "type": "string"
          }
        }
      },
...
      "PaginationResponse": {
        "type": "object",
        "required": [
          "page",
          "per_page",
          "total_items",
          "total_pages"
        ],
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "per_page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total_items": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "total_pages": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "PercentageDiscount": {
        "type": "object",
        "required": [
          "percentage"
        ],
        "properties": {
          "percentage": {
            "type": "string",
            "format": "decimal"
          }
        }
      },
...
      "PlanId": {
        "type": "string",
        "format": "MeteroidId",
        "examples": [
          "plan_7n42DGM5Tflk9n8mt7Fhc7"
        ]
      },
...
      "RestErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "UpdateCouponRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "discount": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CouponDiscountRest"
              }
            ]
          },
          "plan_ids": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/PlanId"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}